feat: AI助手插件与多项优化
- 新增AI助手插件：流式对话、多模型支持、对话历史搜索
- 支持Ollama和OpenAI兼容API（deepseek/doubao/qwen）
- 添加web_search能力判断和智能搜索功能
- 优化导航栏'更多'按钮显示逻辑
- PLUGIN_PRIORITY支持小数排序
- 修复流式显示循环问题，添加'思考中'状态提示

+++ core/plugin_manager.py
@@ -80,7 +80,7 @@ class PluginManager:
                     obj.__module__ == module.__name__):
                     # 获取 PLUGIN_PRIORITY 属性，默认为 999
                     priority = getattr(obj, "PLUGIN_PRIORITY", 999)
-                    return int(priority)
+                    return float(priority)
         except Exception as e:
             self._core.logger.warning(f"Failed to get priority for {plugin_id}: {e}")
